Extract PlayerScanTrophyTitleLoop from ThirtyMinuteCronJob
Peel the per-title trophy scan iteration (fetch, sort, sync, stale-game
cleanup, completion checks) out of ThirtyMinuteCronJob into a dedicated
collaborator, following the existing PlayerScan* extraction pattern.

ThirtyMinuteCronJob now orchestrates login, profile sync, privacy gating,
and finalization while delegating the title loop. Tests moved to
PlayerScanTrophyTitleLoopTest with direct public API access instead of
reflection on the cron orchestrator.
